--- lib/resource/mixins/Saveable.js
@@ -1,34 +1,0 @@
-function applyCustomDataUpdatesIfNecessary(cb){
-  if (!this.customData){
-    return cb();
-  }
-
-  if (this.customData._hasReservedFields()){
-    this.customData = this.customData._deleteReservedFields();
-  }
-
-  if (this.customData._hasRemovedProperties()){
-    return this.customData._deleteRemovedProperties(cb);
-  }
-
-  return cb();
-}
-
-/**
- * Save changes to this resource.
- *
- * @param {Function} callback
- * The function to call when the save operation is complete. Will be called
- * with the parameters (err, updatedResource).
- */
-function saveResource(callback) {
-  var self = this;
-  self._applyCustomDataUpdatesIfNecessary(function () {
-    self.dataStore.saveResource(self, callback);
-  });
-}
-
-module.exports = {
-  save: saveResource,
-  _applyCustomDataUpdatesIfNecessary: applyCustomDataUpdatesIfNecessary
-};
